haveSucceeded() takes a matcher, not a value
`haveSucceeded()` optionally takes a value. If the result is a successful
one, an equality assertion is made using `==`.

This is inflexible--what if I want to assert that a result of type
`Result<[Int], NSError>` contains a specific value?

Change `haveSucceeded()` to take a matcher, not a value. If the result
is successful, the matcher is run against the result's value.

Since `haveSucceeded()` has grown in size, split Guanaco into two files:
one for `haveSucceeded()` and one for `haveFailed()`.